--- gem/templatetags/gem_media_tags.py
@@ -1,17 +1,0 @@
-from django import template
-from molo.core.models import MoloMedia
-
-register = template.Library()
-
-
-@register.inclusion_tag(
-    'gem/tags/media_listing_homepage.html',
-    takes_context=True
-)
-def media_listing_homepage(context):
-    return {
-        'media': MoloMedia.objects.filter(feature_in_homepage=True),
-        'request': context['request'],
-        'is_via_freebasics': context['is_via_freebasics'],
-
-    }
