+++ gem/templatetags/gem_tags.py
@@ -6,6 +6,7 @@ from django.conf import settings
 
 from gem.constants import GENDER
 from gem.models import GemTextBanner
+from molo.core.models import MoloMedia
 from molo.core.templatetags.core_tags import get_pages
 register = Library()
 
@@ -72,3 +73,16 @@ def smart_truncate_chars(value, max_length):
         return truncd_val + '...'
 
     return value
+
+
+@register.inclusion_tag(
+    'gem/tags/media_listing_homepage.html',
+    takes_context=True
+)
+def media_listing_homepage(context):
+    return {
+        'media': MoloMedia.objects.filter(feature_in_homepage=True),
+        'request': context['request'],
+        'is_via_freebasics': context['is_via_freebasics'],
+
+    }
